requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
alembic
python-dotenv
python-jose
celery[redis]
redis
aiobotocore
elasticsearch>=8.0.0
click
pytest
prometheus_client
aioredis
python-multipart
pillow
pyzbar
limits
fastapi-limiter


elasticsearch>=8.0.0
numpy

sentence-transformers